swap out flashcard model
